Move callback_registry tests to their own file
